chore: rename cross-chain to cross-chain-swap and update redirect
